requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
clearbit==0.1.7
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
djangorestframework
djangorestframework-simplejwt
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
psycopg2==2.8.6
pyhunter==1.7
PyJWT==2.0.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
